Ainsley adult hippocampal CA1 dendritic genes

Ainsley adult hippocampal CA1 dendritic genes

AinsleyGenes

Format

character vector of Entrez gene symbols

Source

Ainsley, J.A., Drane, L., Jacobs, J., Kittelberger, K.A., Reijmers, L.G., 2014. Functionally diverse dendritic mRNAs rapidly associate with ribosomes following a novel experience. Nature Communications 5, 4510. doi:10.1038/ncomms5510

Details

Genes derived from Ainsley et al., 2014 mouse adult hippocampal CA1 dendrites.

Gene symbols were updated using Bioconductor package org.Mm.eg.db, using the reported gene_name or Entrez Gene ID from the supplemental data. This step was performed in order to ensure each reference dataset, including the Farris et al. data, were using the same version of the gene symbol annotations.

See also

Examples

data(AinsleyGenes);
## Print the first 50 genes
print(head(AinsleyGenes, 50));
#>  [1] "0610009O20Rik"  "1110059E24Rik"  "1190007I07Rik"  "1600012H06Rik" 
#>  [5] "1700017B05Rik"  "1700123O20Rik"  "2010300C02Rik"  "2510009E07Rik" 
#>  [9] "2700081O15Rik"  "4833439L19Rik"  "4921536K21Rik"  "4930481A15Rik" 
#> [13] "4930522L14Rik"  "4930570G19Rik"  "4933404O12Rik"  "4933417D19Rik" 
#> [17] "4933431E20Rik"  "6330418K02Rik"  "6820431F20Rik"  "9430020K01Rik" 
#> [21] "9930111J21Rik1" "A130040M12Rik"  "A230020J21Rik"  "A330035P11Rik" 
#> [25] "A530040E14Rik"  "Aaas"           "Aagab"          "Aatf"          
#> [29] "Abcb7"          "Abcb10"         "Abcd1"          "Abcf1"         
#> [33] "Abcf2"          "Abcg1"          "Abhd4"          "Abhd8"         
#> [37] "Abhd12"         "Abhd17a"        "Abl1"           "Abl2"          
#> [41] "Ablim1"         "Acadm"          "Acat3"          "Acin1"         
#> [45] "Acly"           "Acp1"           "Acsf2"          "Acsl1"         
#> [49] "Actb"           "Actl6a"