Farris et al gene set enrichment data.frames

Farris et al gene set enrichment data.frames

EnrichmentDFL

Format

An object of class list of length 12.

Source

Farris et al 2019

Details

This object is a list that contains one data.frame for each gene set enrichment test.

Each gene hit list was tested against MSigDB v6.0 using hypergeometric enrichment. The MSigDB data is available from Subramanian, Tamayo, et al. (2005, PNAS 102, 15545-15550) and Liberzon, et al. (2011, Bionformatics)

Examples

data(EnrichmentDFL);

# show the first few rows of one data.frame
head(EnrichmentDFL[[1]])

#>     Comparison      P-value pathGenes geneHits itemsetID Source
#> 1 CA2-CA3_CBup 1.119952e-17      1796      124    M14463    TFT
#> 2 CA2-CA3_CBup 1.242537e-14      2348      140     M5864    TFT
#> 3 CA2-CA3_CBup 1.479887e-12       412       44     M1351    CGP
#> 4 CA2-CA3_CBup 7.532924e-12       667       57    M10961    CGP
#> 5 CA2-CA3_CBup 2.797018e-11       398       41    M15981    CGP
#> 6 CA2-CA3_CBup 5.267156e-09       628       49     M4960    CGP
#>                                     Name
#> 1                         AACTTT_UNKNOWN
#> 2                        CAGGTG_V$E12_Q6
#> 3 SCHAEFFER_PROSTATE_DEVELOPMENT_48HR_DN
#> 4                 GOZGIT_ESR1_TARGETS_DN
#> 5      RIGGI_EWING_SARCOMA_PROGENITOR_UP
#> 6            SMID_BREAST_CANCER_BASAL_DN
#>                                                                                                                                                                                                          Description
#> 1                                                                Genes with promoter regions [-2kb,2kb] around transcription start site containing motif AACTTT. Motif does not match any known transcription factor
#> 2 Genes with promoter regions [-2kb,2kb] around transcription start site containing the motif CAGGTG which matches annotation for TCF3: transcription factor 3 (E2A immunoglobulin enhancer binding factors E12/E47)
#> 3                                                                        Genes down-regulated in the urogenital sinus (UGS) of day E16 females exposed to the androgen dihydrotestosterone [PubChem=10635] for 48 h.
#> 4                                                                       Genes down-regulated in TMX2-28 cells (breast cancer) which do not express ESR1 [GeneID=2099]) compared to the parental MCF7 cells which do.
#> 5                                                                                               Genes up-regulated in mesenchymal stem cells (MSC) engineered to express EWS-FLI1 [GeneID=2130;2321] fusion protein.
#> 6                                                                                                                                                     Genes down-regulated in basal subtype of breast cancer samles.
#>   Category
#> 1       C3
#> 2       C3
#> 3       C2
#> 4       C2
#> 5       C2
#> 6       C2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                          geneNames
#> 1                                                                                                             Acan,Ackr3,Adam9,Adcy6,Adgrb3,Aldh1a1,Amigo2,Ap1s2,Apbb2,Arap2,Arhgap44,Arpc2,Avpr1b,Bcl11b,Bcl6,Bhlhe22,Cacna1h,Cacna2d3,Cdkl5,Chst1,Cpne8,Crim1,Csrnp3,Dab1,Dock11,Dtna,Ephx2,Etv1,Fam126a,Fam19a2,Fam49a,Fam81a,Farp1,Fezf2,Fgf12,Fgf16,Fgfr1,Gabra3,Gabrb1,Gng4,Hcfc1r1,Hcn1,Hcrtr2,Hectd2,Hrk,Igf1r,Il16,Ildr2,Kcnj3,Kcnn2,Laptm4b,Large1,Lhx2,Lrp1b,Lrrc8d,Lrrtm4,Maml3,Map4k3,Mcu,Meis2,Mid2,Mitf,Mkl1,Mmp16,Nectin1,Nfia,Nmt2,Nptx2,Nr3c2,Ntm,Nudt4,Oxct1,Pcdh10,Pde1a,Pde4d,Pex5l,Pfn1,Pfn2,Pias1,Pitpnc1,Plpp6,Plxnc1,Pmepa1,Pou3f1,Ppm1e,Ppp2r2b,Pygo1,Rab3c,Rap2b,Rapgef5,Rfx3,Rnf128,Scrt2,Sel1l3,Sh3kbp1,Slain1,Slc24a3,Slc24a4,Slc4a7,Slc6a11,Slc6a15,Slitrk3,Slitrk5,Smad1,Sostdc1,Sox5,Ssh2,St8sia1,Stard13,Stc1,Stk38l,Supt6,Tcf4,Tle1,Tox2,Trib1,Trim2,Trpm3,Tubb2b,Usp13,Vit,Wasl,Zbtb16,Zfp804a
#> 2 Ablim3,Abr,Ackr3,Acvr2a,Adam10,Aldh1a1,Angpt1,Ap1s2,Aph1b,Arap2,Arhgap26,Arhgap44,Asic2,Bcl11b,Bcl6,Bhlhe22,Bzw2,C1qtnf4,Cacna1h,Cacna2d3,Camta2,Car7,Cast,Coro1c,Crim1,Dapk1,Dchs2,Dgkg,Dmxl1,Doc2b,Drd5,Dscam,Dstn,Dusp4,Dusp5,Erc1,Etv1,Extl3,Fabp3,Fam126a,Fam19a2,Fam49a,Fam81a,Fezf2,Fgf12,Fgf5,Fkbp5,Gabra3,Gabrb1,Gabrg2,Gad1,Gmpr,Gna12,Gpd1,Hcrtr2,Hdac9,Hectd2,Igf1r,Ildr2,Kcnd2,Kcnh7,Kcnj4,Kcnn2,Lmo2,Lrig1,Lrrc8c,Lrrc8d,Map3k19,Map4k3,Meis2,Mfsd4a,Mical2,Mid2,Mipep,Mkl1,Mmp16,Mylk,Nectin1,Nedd4l,Negr1,Nmt2,Npr3,Nptx2,Nr3c2,Nrip3,Nrxn3,Nt5dc3,Ntf3,Osbpl10,Pak6,Palld,Pcdh1,Pcp4,Pianp,Pias1,Pigt,Pip5k1b,Plxnc1,Pmepa1,Pou3f1,Ppp2r2b,Ptprj,Rapgef5,Rfx3,Rhbdl1,Rnf128,Scg5,Scrt2,Sel1l3,Sh3bgrl2,Sh3rf1,Slc1a1,Slc24a3,Slc25a5,Slc4a8,Slc6a11,Slitrk3,Sms,Snx25,Sostdc1,Sowaha,Srgap2,Ssh2,Stard13,Stc1,Supt6,Tbc1d8,Tdrp,Tle1,Tmem178,Tmem28,Tmsb10,Tox2,Trib1,Trim37,Tspan13,Ubxn2a,Wasl,Zbtb16,Zdhhc15
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    Akap5,Ar,Arid5b,Cacna2d3,Ccdc88c,Chrna4,Dubr,Dusp4,Elovl6,Erc1,Fam126a,Fam131a,Fam46a,Fam49a,Farp1,Gabrb1,Hunk,Isoc1,Itga4,Laptm4b,Large1,Lin7a,Matn2,Mcur1,Mex3b,Mmp16,Msrb2,Negr1,Nuak1,Oxct1,Palld,Pcdh10,Pde1a,Pde1b,Pex5l,Pmepa1,Rhbdl3,Sh3kbp1,Shisa2,Sipa1l2,Slc1a1,Tcf4,Tgfb1i1,Tiam2
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   Abat,Abhd2,Ablim1,Ackr3,Ahcyl2,Akap5,Apbb2,Ar,Arg2,Arid5b,Atrnl1,Bcl6,Cap2,Dclk1,Dscam,Dtna,Fam46a,Farp1,Gad1,Hdac9,Htr7,Igf1r,Kcnj3,Large1,Lin7a,Mid2,Mitf,Mmp16,Nab1,Ninj1,Nuak1,Pcdh10,Pcp4,Pfn2,Pgm2l1,Pmepa1,Ppm1e,Prkca,Prss23,Rara,Rassf5,Runx2,Sec14l1,Sh3bgrl2,Sipa1l2,Slc1a1,Slc24a3,Stc1,Tcf4,Thsd4,Tle1,Tmtc1,Trim2,Trim37,Tspan13,Zc2hc1a,Zdhhc15
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Aldh5a1,Arhgef6,Bcl11b,Bhlhe22,Cadps2,Crip2,Csrnp3,Dab1,Dapk1,Dusp5,Efhd2,F2r,Farp1,Fat3,Fat4,Hdac9,Hist1h2bc,Hs6st1,Itga4,Kalrn,Lmo2,Mcur1,Npr3,Nptx2,Npy1r,Nr3c2,Olfm3,Pgm2l1,Rbm24,Reep1,Sec14l1,Sel1l3,Shisa9,Slain1,Slc24a3,Tcf4,Tox2,Tspan13,Wdfy2,Zbtb16,Zfp703
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                            Aamdc,Abat,Ablim3,Ackr3,Aldh1a1,Amigo2,Apbb2,Ar,Asph,Atrnl1,Brinp2,Cacna1h,Cacna2d3,Chst1,Crip2,Dclk1,Dusp4,Dusp5,Ephx2,Fgfr1,Gpd1,Hcfc1r1,Hspa2,Igf1r,Inpp5j,Kcnj3,Lin7a,Lrig1,Mipep,Nedd4l,Nkain1,Npy1r,Nudt4,Prss23,Ptprn2,Qdpr,Rara,Reep1,Rhbdl1,Slc1a1,Slc38a1,Slc4a7,Slc4a8,Sorl1,Sri,Supt6,Tspan13,Ubl3,Zbtb16

# If you have the tibble package, it has a clean display
if (require(tibble)) {
   print(tibble::tibble(EnrichmentDFL[[1]]))
}

#> Loading required package: tibble

#> # A tibble: 3,539 x 1
#>    `EnrichmentDFL[… $`P-value` $pathGenes $geneHits $itemsetID $Source $Name
#>    <chr>                 <dbl>      <dbl>     <dbl> <chr>      <chr>   <chr>
#>  1 CA2-CA3_CBup       1.12e-17       1796       124 M14463     TFT     AACT…
#>  2 CA2-CA3_CBup       1.24e-14       2348       140 M5864      TFT     CAGG…
#>  3 CA2-CA3_CBup       1.48e-12        412        44 M1351      CGP     SCHA…
#>  4 CA2-CA3_CBup       7.53e-12        667        57 M10961     CGP     GOZG…
#>  5 CA2-CA3_CBup       2.80e-11        398        41 M15981     CGP     RIGG…
#>  6 CA2-CA3_CBup       5.27e- 9        628        49 M4960      CGP     SMID…
#>  7 CA2-CA3_CBup       1.21e- 8        273        29 M735       CP:REA… REAC…
#>  8 CA2-CA3_CBup       3.55e- 8        411        36 M1954      CGP     MIKK…
#>  9 CA2-CA3_CBup       3.61e- 8        190        23 M2817      CGP     STAR…
#> 10 CA2-CA3_CBup       6.31e- 8       1005        64 M10371     CGP     BENP…
#> # … with 3,529 more rows, and 3 more variables: $Description <chr>,
#> #   $Category <chr>, $geneNames <chr>

# the jamba package sdim() is a good summary
jamba::sdim(EnrichmentDFL);

#>              rows cols      class
#> CA2-CA3_CBup 3539   10 data.frame
#> CA2-CA1_CBup 5241   10 data.frame
#> CA2-DG_CBup  6737   10 data.frame
#> CA2-CA3_DEup  832   10 data.frame
#> CA2-CA1_DEup 4168   10 data.frame
#> CA2-DG_DEup  2424   10 data.frame
#> CA2-CA3_CBdn 3160   10 data.frame
#> CA2-CA1_CBdn 3664   10 data.frame
#> CA2-DG_CBdn  5062   10 data.frame
#> CA2-CA3_DEdn 1035   10 data.frame
#> CA2-CA1_DEdn 4031   10 data.frame
#> CA2-DG_DEdn  2797   10 data.frame