Nakayama adult hippocampal CA1 dendritic genes

Nakayama adult hippocampal CA1 dendritic genes

NakayamaGenes

Format

character vector of Entrez gene symbols

Source

Nakayama, K., Ohashi, R., Shinoda, Y., Yamazaki, M., Abe, M., Fujikawa, A., Shigenobu, S., Futatsugi, A., Noda, M., Mikoshiba, K., Furuichi, T., Sakimura, K., Shiina, N., 2017. RNG105/caprin1, an RNA granule protein for dendritic mRNA localization, is essential for long-term memory formation. eLife 6, e29677. doi:10.7554/eLife.29677

Details

Genes derived from Nakayama et al., 2017 mouse adult hippocampal CA1 dendrites.

Gene symbols were updated using Bioconductor package org.Mm.eg.db, using the reported gene_name or Entrez Gene ID from the supplemental data. This step was performed in order to ensure each reference dataset, including the Farris et al. data, were using the same version of the gene symbol annotations.

See also

Examples

data(NakayamaGenes);
## Print the first 50 genes
print(head(NakayamaGenes, 50));

#>  [1] "1700001O22Rik"  "1810011H11Rik"  "1810012K08Rik"  "2300009A05Rik" 
#>  [5] "4932438H23Rik"  "6430550D23Rik"  "9330132A10Rik"  "9930111J21Rik2"
#>  [9] "A430105I19Rik"  "A630001G21Rik"  "Abat"           "Abcc3"         
#> [13] "Abhd17a"        "Acad12"         "Acap3"          "Accs"          
#> [17] "Acer2"          "Acox3"          "Acsf2"          "Adamts10"      
#> [21] "Adamtsl5"       "Adap1"          "Adap2"          "Adcy1"         
#> [25] "Adcy7"          "Adora3"         "Adra1b"         "Adra2c"        
#> [29] "Adrb2"          "Aen"            "Afap1"          "Afap1l2"       
#> [33] "Agap1"          "Agap2"          "Agap3"          "Agmo"          
#> [37] "Agpat2"         "Ahcyl2"         "Aif1"           "Aif1l"         
#> [41] "Akip1"          "Akr1b10"        "Aldh1l2"        "Alkbh5"        
#> [45] "Alkbh6"         "Alpk1"          "Alpl"           "Amer2"         
#> [49] "Amt"            "Anapc13"