Farris et al gene-level expression data

Farris et al gene-level expression data

farrisGeneSE

Format

SummarizedExperiment object with 49341 gene rows, and 24 sample columns. Sample annotations are accessed as described for SummarizedExperiment-class objects, using SummarizedExperiment::colData(). Gene annotations are accessed using SummarizedExperiment::rowData().

The colData colnames are described as follows:

CellType

The major excitatory pyramidal cell types in the mouse hippocampus, CA1, CA2, CA3, and dentate gyrus (DG).

Compartment

The cellular compartment within each cell type: CB (cell body) is the pyramidal cell layer; DE (dendrite) is stratum radiatum of CA1, CA2, CA3, and the molecular layer of DG.

AnimalID

The number assigned to each mouse.

groupName

The sample group name associated with each CellType and Compartment.

The count data is accessed through the function assays and has the following entries:

counts

Median-normalized log2-transformed pseudocount data per-gene generated by Salmon, after import using tximport::tximport(). The log2 transformation was log2(1+x).

raw_counts

The raw log2-transformed pseudocount data per-gene generated by Salmon. The log2 transformation was log2(1+x).

The rowData contains gene annotations, in this case limited to the Entrez gene symbol.

Source

Farris et al publication.

Details

SummarizedExperiment object containing gene-level expression data following Salmon quantitation, and import using tximport::tximport() at the gene level. Data is stored as log2-transformed pseudocounts from Salmon, after median normalization.

Examples

data(farrisGeneSE);
## Gene annotation DataFrame
head(SummarizedExperiment::rowData(farrisGeneSE))

#> DataFrame with 6 rows and 4 columns
#>                       probeID      ProbeName       GeneName     geneSymbol
#>                   <character>    <character>    <character>    <character>
#> -343C11.2           -343C11.2      -343C11.2      -343C11.2      -343C11.2
#> 00R_AC107638.2 00R_AC107638.2 00R_AC107638.2 00R_AC107638.2 00R_AC107638.2
#> 00R_Pgap2           00R_Pgap2      00R_Pgap2      00R_Pgap2      00R_Pgap2
#> 0610005C13Rik   0610005C13Rik  0610005C13Rik  0610005C13Rik  0610005C13Rik
#> 0610006L08Rik   0610006L08Rik  0610006L08Rik  0610006L08Rik  0610006L08Rik
#> 0610007P14Rik   0610007P14Rik  0610007P14Rik  0610007P14Rik  0610007P14Rik

## Sample annotation DataFrame
head(SummarizedExperiment::colData(farrisGeneSE))

#> DataFrame with 6 rows and 4 columns
#>          CellType Compartment    AnimalID groupName
#>          <factor>    <factor> <character>  <factor>
#> CA2CB492      CA2          CB         492    CA2_CB
#> CA2CB496      CA2          CB         496    CA2_CB
#> CA2CB502      CA2          CB         502    CA2_CB
#> CA2DE492      CA2          DE         492    CA2_DE
#> CA2DE496      CA2          DE         496    CA2_DE
#> CA2DE502      CA2          DE         502    CA2_DE

## Assay data matrix for normalized counts
head(SummarizedExperiment::assays(farrisGeneSE)[["counts"]]);

#>                 CA2CB492   CA2CB496   CA2CB502 CA2DE492 CA2DE496 CA2DE502
#> -343C11.2       5.125592  5.5075194  6.7499062 7.892555 6.632026 4.392167
#> 00R_AC107638.2  0.000000  0.0000000  0.3814014 4.789433 3.625500 4.577885
#> 00R_Pgap2       0.000000  0.2614108  0.3792121 0.000000 0.000000 0.000000
#> 0610005C13Rik   3.733125  1.5009798  0.4412405 1.131151 4.405658 5.000691
#> 0610006L08Rik   0.000000  0.0000000  0.0000000 0.000000 0.000000 0.000000
#> 0610007P14Rik  10.151877 10.2983772 10.8568017 9.610023 9.658824 9.517425
#>                 CA1CB492  CA1CB496    CA1CB502  CA1DE492  CA1DE496  CA1DE502
#> -343C11.2       6.912525  5.735207  6.05669795  5.393926 3.9618802 2.1088010
#> 00R_AC107638.2  0.000000  0.000000  3.12930078  0.000000 0.0000000 0.0000000
#> 00R_Pgap2       0.000000  0.000000  0.02389523  0.000000 0.3958771 0.2753541
#> 0610005C13Rik   1.599389  4.439576  2.70020462  3.158550 0.9273044 6.2195594
#> 0610006L08Rik   0.000000  0.000000  0.00000000  0.000000 0.0000000 0.0000000
#> 0610007P14Rik  10.465095 10.094485 10.04557169 10.081736 9.4668697 9.6611932
#>                  CA3CB492  CA3CB496  CA3CB502 CA3DE492  CA3DE496 CA3DE502
#> -343C11.2       5.9012222  5.865129  6.233813 7.230335 7.6758413 5.081336
#> 00R_AC107638.2  0.0000000  0.000000  0.000000 6.925011 0.1730263 0.000000
#> 00R_Pgap2       0.2260925  0.000000  0.429772 5.221890 0.2015875 0.000000
#> 0610005C13Rik   0.2042486  4.011457  3.399649 7.013474 2.3532787 1.236412
#> 0610006L08Rik   0.0000000  0.000000  0.000000 0.000000 0.0000000 0.000000
#> 0610007P14Rik  10.5046539 10.241662 10.283002 9.552535 9.8617314 9.791313
#>                 DGCB492  DGCB496  DGCB502  DGDE492    DGDE496  DGDE502
#> -343C11.2      6.126416 6.635583 5.785136 6.589268 2.87045723 4.544814
#> 00R_AC107638.2 0.000000 2.113397 0.000000 0.000000 0.00000000 3.614111
#> 00R_Pgap2      0.000000 0.000000 0.000000 0.000000 0.02607423 0.000000
#> 0610005C13Rik  5.356292 4.303088 4.565700 2.431196 1.37349373 2.947646
#> 0610006L08Rik  0.000000 0.000000 0.000000 0.000000 0.00000000 0.000000
#> 0610007P14Rik  9.897466 9.985364 9.992884 8.695595 9.63679389 8.899735