Sashimi plot files data.frame for Farris data

Sashimi plot files data.frame for Farris data

farris_sashimi_files_df

Format

data.frame suitable for use with splicejam::prepareSashimi() for the filesDF argument. The columns:

"sample_id" - sample label used for each panel in the Sashimi plot

"url" - full web URL or file path (relative or absolute, from the working directory of the R-shiny app) to each file. It should contain .bed.gz files for junctions, and big wig files for coverage data, using .pos.bw for positive strand coverage, and .neg.bw for negative strand coverage, where applicable.

"type" - character string indicating the type of data, using "junction" or "bw".

"scale_factor" - optional column with numeric value to adjust the coverage data value, by multiplying the scale factor. The default scale factor is 1.

"CellType","Compartment" - additional columns are used in the sample selection HTML widget to help organize and sort samples as appropriate.

Examples

#farris_sashimi_files_df
base_url <- "https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/";
factor1 <- c("CA1", "CA2", "CA3", "DG");
factor2 <- c("CB", "DE");
junc_suffix <- ".STAR_mm10.combinedJunctions.bed.gz";
junc_urls <- paste0(base_url,
   rep(factor1, each=2),
   "_",
   rep(factor2, 4),
   junc_suffix);
junc_sample_id <- paste(rep(factor1, each=2),
   rep(factor2, 4),
   sep="_");
junc_scale_factors <- 2^c(
   CA1_CB=-0.185,
   CA1_DE=0.146,
   CA2_CB=-0.241,
   CA2_DE=0.765,
   CA3_CB=-0.215,
   CA3_DE=0.256,
   DG_CB=-0.376,
   DG_DE=0.372);
junc_df <- data.frame(sample_id=junc_sample_id,
   url=junc_urls,
   type="junction",
   scale_factor=junc_scale_factors[junc_sample_id]);
if (suppressPackageStartupMessages(require(knitr))) {
   knitr::kable(junc_df)
} else {
   junc_df;
}
#> 
#> 
#> |       |sample_id |url                                                                                             |type     | scale_factor|
#> |:------|:---------|:-----------------------------------------------------------------------------------------------|:--------|------------:|
#> |CA1_CB |CA1_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA1_CB.STAR_mm10.combinedJunctions.bed.gz |junction |    0.8796491|
#> |CA1_DE |CA1_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA1_DE.STAR_mm10.combinedJunctions.bed.gz |junction |    1.1064974|
#> |CA2_CB |CA2_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA2_CB.STAR_mm10.combinedJunctions.bed.gz |junction |    0.8461586|
#> |CA2_DE |CA2_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA2_DE.STAR_mm10.combinedJunctions.bed.gz |junction |    1.6993700|
#> |CA3_CB |CA3_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA3_CB.STAR_mm10.combinedJunctions.bed.gz |junction |    0.8615462|
#> |CA3_DE |CA3_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA3_DE.STAR_mm10.combinedJunctions.bed.gz |junction |    1.1941632|
#> |DG_CB  |DG_CB     |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/DG_CB.STAR_mm10.combinedJunctions.bed.gz  |junction |    0.7705711|
#> |DG_DE  |DG_DE     |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/DG_DE.STAR_mm10.combinedJunctions.bed.gz  |junction |    1.2941457|

bw_factor1 <- c(factor1);
bw_factor2 <- c(factor2);
bw_base_url <- paste0(base_url, "union_bigwig/");
bw_strand <- c("pos", "neg");
bw_suffix <- ".bw";
bw_urls <- paste0(
   rep(bw_factor1, each=4),
   "_",
   rep(bw_factor2, each=2),
   ".union.",
   bw_strand,
   bw_suffix
);
bw_scale_factors <- 2^c(
   CA1_CB=-0.0157,
   CA1_DE=0.306,
   CA2_CB=-0.0266,
   CA2_DE=0.0947,
   CA3_CB=0.00839,
   CA3_DE=0.528,
   DG_CB=-0.243,
   DG_DE=-0.283);

if (suppressPackageStartupMessages(require(knitr))) {
   knitr::kable(as.data.frame(matrix(bw_urls, ncol=5)))
} else {
   as.data.frame(matrix(bw_urls, ncol=4));
}
#> Warning: data length [16] is not a sub-multiple or multiple of the number of columns [5]
#> 
#> 
#> |V1                  |V2                  |V3                  |V4                 |V5                  |
#> |:-------------------|:-------------------|:-------------------|:------------------|:-------------------|
#> |CA1_CB.union.pos.bw |CA2_CB.union.pos.bw |CA3_CB.union.pos.bw |DG_CB.union.pos.bw |CA1_CB.union.pos.bw |
#> |CA1_CB.union.neg.bw |CA2_CB.union.neg.bw |CA3_CB.union.neg.bw |DG_CB.union.neg.bw |CA1_CB.union.neg.bw |
#> |CA1_DE.union.pos.bw |CA2_DE.union.pos.bw |CA3_DE.union.pos.bw |DG_DE.union.pos.bw |CA1_DE.union.pos.bw |
#> |CA1_DE.union.neg.bw |CA2_DE.union.neg.bw |CA3_DE.union.neg.bw |DG_DE.union.neg.bw |CA1_DE.union.neg.bw |

# remove NA entries
bw_url <- jamba::rmNA(bw_urls);
bw_sample_id <- gsub("^.*(CA[123]|DG)[_]*(CB|DE).*",
   "\\1_\\2",
   bw_url);
bw_df <- data.frame(
   url=paste0(bw_base_url, bw_url),
   sample_id=bw_sample_id,
   type="bw",
   scale_factor=bw_scale_factors[bw_sample_id]);
bw_df <- bw_df[do.call(order, bw_df[,c(2,1,3)]),];

farris_sashimi_files_df <- rbind(
   junc_df,
   bw_df);
# replace missing scale_factor with 1, should not be necessary
farris_sashimi_files_df$scale_factor[is.na(farris_sashimi_files_df$scale_factor)] <- 1;
farris_sashimi_files_df$CellType <- factor(
   gsub("_.+", "",
      farris_sashimi_files_df$sample_id),
   levels=factor1);
farris_sashimi_files_df$Compartment <- factor(
   gsub("^.+_", "",
      farris_sashimi_files_df$sample_id),
   levels=factor2);
if (suppressPackageStartupMessages(require(knitr))) {
   knitr::kable(farris_sashimi_files_df)
} else {
   farris_sashimi_files_df;
}
#> 
#> 
#> |       |sample_id |url                                                                                             |type     | scale_factor|CellType |Compartment |
#> |:------|:---------|:-----------------------------------------------------------------------------------------------|:--------|------------:|:--------|:-----------|
#> |CA1_CB |CA1_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA1_CB.STAR_mm10.combinedJunctions.bed.gz |junction |    0.8796491|CA1      |CB          |
#> |CA1_DE |CA1_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA1_DE.STAR_mm10.combinedJunctions.bed.gz |junction |    1.1064974|CA1      |DE          |
#> |CA2_CB |CA2_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA2_CB.STAR_mm10.combinedJunctions.bed.gz |junction |    0.8461586|CA2      |CB          |
#> |CA2_DE |CA2_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA2_DE.STAR_mm10.combinedJunctions.bed.gz |junction |    1.6993700|CA2      |DE          |
#> |CA3_CB |CA3_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA3_CB.STAR_mm10.combinedJunctions.bed.gz |junction |    0.8615462|CA3      |CB          |
#> |CA3_DE |CA3_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/CA3_DE.STAR_mm10.combinedJunctions.bed.gz |junction |    1.1941632|CA3      |DE          |
#> |DG_CB  |DG_CB     |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/DG_CB.STAR_mm10.combinedJunctions.bed.gz  |junction |    0.7705711|DG       |CB          |
#> |DG_DE  |DG_DE     |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/DG_DE.STAR_mm10.combinedJunctions.bed.gz  |junction |    1.2941457|DG       |DE          |
#> |2      |CA1_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA1_CB.union.neg.bw          |bw       |    0.9891766|CA1      |CB          |
#> |1      |CA1_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA1_CB.union.pos.bw          |bw       |    0.9891766|CA1      |CB          |
#> |4      |CA1_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA1_DE.union.neg.bw          |bw       |    1.2362753|CA1      |DE          |
#> |3      |CA1_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA1_DE.union.pos.bw          |bw       |    1.2362753|CA1      |DE          |
#> |6      |CA2_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA2_CB.union.neg.bw          |bw       |    0.9817312|CA2      |CB          |
#> |5      |CA2_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA2_CB.union.pos.bw          |bw       |    0.9817312|CA2      |CB          |
#> |8      |CA2_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA2_DE.union.neg.bw          |bw       |    1.0678433|CA2      |DE          |
#> |7      |CA2_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA2_DE.union.pos.bw          |bw       |    1.0678433|CA2      |DE          |
#> |10     |CA3_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA3_CB.union.neg.bw          |bw       |    1.0058324|CA3      |CB          |
#> |9      |CA3_CB    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA3_CB.union.pos.bw          |bw       |    1.0058324|CA3      |CB          |
#> |12     |CA3_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA3_DE.union.neg.bw          |bw       |    1.4419289|CA3      |DE          |
#> |11     |CA3_DE    |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/CA3_DE.union.pos.bw          |bw       |    1.4419289|CA3      |DE          |
#> |14     |DG_CB     |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/DG_CB.union.neg.bw           |bw       |    0.8449864|DG       |CB          |
#> |13     |DG_CB     |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/DG_CB.union.pos.bw           |bw       |    0.8449864|DG       |CB          |
#> |16     |DG_DE     |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/DG_DE.union.neg.bw           |bw       |    0.8218802|DG       |DE          |
#> |15     |DG_DE     |https://snpinfo.niehs.nih.gov/ucscview/farrisHub/mm10/union_bigwig/DG_DE.union.pos.bw           |bw       |    0.8218802|DG       |DE          |

table(farris_sashimi_files_df[,c("CellType", "Compartment", "type")]);
#> , , type = bw
#> 
#>         Compartment
#> CellType CB DE
#>      CA1  2  2
#>      CA2  2  2
#>      CA3  2  2
#>      DG   2  2
#> 
#> , , type = junction
#> 
#>         Compartment
#> CellType CB DE
#>      CA1  1  1
#>      CA2  1  1
#>      CA3  1  1
#>      DG   1  1
#>