farrisdata: Farris et al RNA-seq data

Supporting data for Farris et al Mouse hippocampus RNA-seq analysis workflow, utilizing the jampack package.

Details

These objects are provided as SummarizedExperiment objects as described in the Bioconductor package SummarizedExperiment.

Available data

farrisGeneSE, farrisTxSE

Gene and transcript models

Gene and transcript models were defined in the corresponding Gencode GTF file available from the Gencode FTP site:

gencode.vM12.annotation.gtf.gz

Alignment to the transcriptome was performed using the FASTA sequence file also from Gencode FTP site:

gencode.vM12.transcripts.fa.gz

Gencode data is subject to Creative Commons License (Attribution, Non-commercial, No-derivatives 2.5 license), see https://creativecommons.org/licenses/by-nc-nd/2.5/.