Cajigas rat hippocampal CA1 dendritic genes

Cajigas rat hippocampal CA1 dendritic genes

CajigasGenes

Format

character vector of Entrez gene symbols

Source

Cajigas, I.J., Tushev, G., Will, T.J., tom Dieck, S., Fuerst, N., Schuman, E.M., 2012. The local transcriptome in the synaptic neuropil revealed by deep sequencing and high-resolution imag-ing. Neuron 74, 453–466. doi:10.1016/j.neuron.2012.02.036

Details

Genes derived from Cajigas et al., 2012 rat hippocampal CA1 dendrites.

Gene symbols were updated using Bioconductor package org.Rn.eg.db, using the reported gene_name or Entrez Gene ID from the supplemental data. This step was performed in order to ensure each reference dataset, including the Farris et al. data, were using the same version of the gene symbol annotations.

See also

Examples

data(CajigasGenes);
## Print the first 50 genes
print(head(CajigasGenes, 50));
#>  [1] "Aaas"    "Aadacl1" "Aars"    "Aarsd1"  "Abca3"   "Abcc8"   "Abcf1"  
#>  [8] "Abhd2"   "Abhd12"  "Abhd16a" "Abi1"    "Ablim3"  "Acap3"   "Acbd6"  
#> [15] "Accn1"   "Aco2"    "Acot7"   "Acpl2"   "Acta1"   "Actb"    "Actg1"  
#> [22] "Actn1"   "Actr1b"  "Actr2"   "Actr3"   "Actr3b"  "Acvr2a"  "Adam10" 
#> [29] "Adam11"  "Adam23"  "Adamts3" "Adap1"   "Adap2"   "Adcy1"   "Adcy3"  
#> [36] "Adcy5"   "Adcy9"   "Adcyap1" "Add2"    "Adora1"  "Adora3"  "Adprh"  
#> [43] "Adrm1"   "Adss"    "Aff2"    "Aff3"    "Afmid"   "Agap3"   "Aggf1"  
#> [50] "Agpat1"