Farris et al gene-level expression data

Farris et al gene-level expression data

farrisGeneSE

Format

SummarizedExperiment object with 49341 gene rows, and 24 sample columns. Sample annotations are accessed as described for SummarizedExperiment-class objects, using SummarizedExperiment::colData(). Gene annotations are accessed using SummarizedExperiment::rowData().

The colData colnames are described as follows:

CellType

The major excitatory pyramidal cell types in the mouse hippocampus, CA1, CA2, CA3, and dentate gyrus (DG).

Compartment

The cellular compartment within each cell type: CB (cell body) is the pyramidal cell layer; DE (dendrite) is stratum radiatum of CA1, CA2, CA3, and the molecular layer of DG.

AnimalID

The number assigned to each mouse.

groupName

The sample group name associated with each CellType and Compartment.

The count data is accessed through the function assays and has the following entries:

counts

Median-normalized log2-transformed pseudocount data per-gene generated by Salmon, after import using tximport::tximport(). The log2 transformation was log2(1+x).

raw_counts

The raw log2-transformed pseudocount data per-gene generated by Salmon. The log2 transformation was log2(1+x).

The rowData contains gene annotations, in this case limited to the Entrez gene symbol.

Source

Farris et al publication.

Details

SummarizedExperiment object containing gene-level expression data following Salmon quantitation, and import using tximport::tximport() at the gene level. Data is stored as log2-transformed pseudocounts from Salmon, after median normalization.

See also

Examples

data(farrisGeneSE);
## Gene annotation DataFrame
if (suppressPackageStartupMessages(require(knitr))) {
   knitr::kable(
      tail(head(SummarizedExperiment::rowData(farrisGeneSE), 3079), 10)
   )
} else {
   tail(head(SummarizedExperiment::rowData(farrisGeneSE), 3079), 10)
}
#> 
#> 
#> |         |probeID  |ProbeName |GeneName |geneSymbol |
#> |:--------|:--------|:---------|:--------|:----------|
#> |Amigo1   |Amigo1   |Amigo1    |Amigo1   |Amigo1     |
#> |Amigo2   |Amigo2   |Amigo2    |Amigo2   |Amigo2     |
#> |Amigo3   |Amigo3   |Amigo3    |Amigo3   |Amigo3     |
#> |Ammecr1  |Ammecr1  |Ammecr1   |Ammecr1  |Ammecr1    |
#> |Ammecr1l |Ammecr1l |Ammecr1l  |Ammecr1l |Ammecr1l   |
#> |Amn      |Amn      |Amn       |Amn      |Amn        |
#> |Amn1     |Amn1     |Amn1      |Amn1     |Amn1       |
#> |Amot     |Amot     |Amot      |Amot     |Amot       |
#> |Amotl1   |Amotl1   |Amotl1    |Amotl1   |Amotl1     |
#> |Amotl2   |Amotl2   |Amotl2    |Amotl2   |Amotl2     |

## Sample annotation DataFrame
if (suppressPackageStartupMessages(require(knitr))) {
   knitr::kable(
      SummarizedExperiment::colData(farrisGeneSE)
   )
} else {
   SummarizedExperiment::colData(farrisGeneSE)
}
#> 
#> 
#> |         |CellType |Compartment |AnimalID |groupName |
#> |:--------|:--------|:-----------|:--------|:---------|
#> |CA2CB492 |CA2      |CB          |492      |CA2_CB    |
#> |CA2CB496 |CA2      |CB          |496      |CA2_CB    |
#> |CA2CB502 |CA2      |CB          |502      |CA2_CB    |
#> |CA2DE492 |CA2      |DE          |492      |CA2_DE    |
#> |CA2DE496 |CA2      |DE          |496      |CA2_DE    |
#> |CA2DE502 |CA2      |DE          |502      |CA2_DE    |
#> |CA1CB492 |CA1      |CB          |492      |CA1_CB    |
#> |CA1CB496 |CA1      |CB          |496      |CA1_CB    |
#> |CA1CB502 |CA1      |CB          |502      |CA1_CB    |
#> |CA1DE492 |CA1      |DE          |492      |CA1_DE    |
#> |CA1DE496 |CA1      |DE          |496      |CA1_DE    |
#> |CA1DE502 |CA1      |DE          |502      |CA1_DE    |
#> |CA3CB492 |CA3      |CB          |492      |CA3_CB    |
#> |CA3CB496 |CA3      |CB          |496      |CA3_CB    |
#> |CA3CB502 |CA3      |CB          |502      |CA3_CB    |
#> |CA3DE492 |CA3      |DE          |492      |CA3_DE    |
#> |CA3DE496 |CA3      |DE          |496      |CA3_DE    |
#> |CA3DE502 |CA3      |DE          |502      |CA3_DE    |
#> |DGCB492  |DG       |CB          |492      |DG_CB     |
#> |DGCB496  |DG       |CB          |496      |DG_CB     |
#> |DGCB502  |DG       |CB          |502      |DG_CB     |
#> |DGDE492  |DG       |DE          |492      |DG_DE     |
#> |DGDE496  |DG       |DE          |496      |DG_DE     |
#> |DGDE502  |DG       |DE          |502      |DG_DE     |

## Assay data matrix for normalized counts
if (suppressPackageStartupMessages(require(knitr))) {
   knitr::kable(
      tail(head(SummarizedExperiment::assays(farrisGeneSE)[["counts"]], 3079), 10)
   )
} else {
   tail(head(SummarizedExperiment::assays(farrisGeneSE)[["counts"]], 3079), 10)
}
#> 
#> 
#> |         |  CA2CB492|  CA2CB496|  CA2CB502|  CA2DE492|  CA2DE496|  CA2DE502|  CA1CB492|  CA1CB496|  CA1CB502|  CA1DE492|  CA1DE496|  CA1DE502|  CA3CB492|  CA3CB496|  CA3CB502|  CA3DE492|  CA3DE496|  CA3DE502|   DGCB492|   DGCB496|   DGCB502|   DGDE492|   DGDE496|   DGDE502|
#> |:--------|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|
#> |Amigo1   | 11.788221| 11.911877| 11.932830|  9.569232| 10.573725| 10.256418| 11.752525| 11.214424| 11.312252| 11.475075| 10.639000|  9.610328| 11.795311| 11.561756| 11.723912|  9.967730| 11.514411| 10.071652| 11.515426| 11.459975| 11.529161|  9.946452| 11.219920|  9.750555|
#> |Amigo2   | 14.289255| 14.058009| 14.357035| 10.189114|  9.764771|  9.527168|  7.812238|  7.241192|  8.526366|  8.492178|  9.472811|  9.132392|  9.826938|  9.756097|  9.568759|  8.737075|  8.966104|  9.600518|  8.313740|  8.417530|  8.347074|  9.467638|  8.668333|  9.076533|
#> |Amigo3   |  6.872062|  6.536729|  5.191612|  6.957927|  0.000000|  6.080202|  6.884087|  6.247739|  6.482084|  0.000000|  6.248117|  6.704929|  5.498212|  5.692095|  6.587664|  0.000000|  6.595790|  0.000000|  6.616847|  5.911659|  5.595349|  5.423117|  0.000000|  5.571091|
#> |Ammecr1  |  6.260452|  6.416658|  6.754910|  6.034518|  7.586801|  7.275456|  6.649984|  6.764007|  7.346509|  0.000000|  9.331435|  6.891055|  7.212692|  5.866938|  6.663303|  8.729234|  8.282891|  7.485690|  6.755224|  7.414521|  6.711571|  7.151134|  8.957837|  7.311669|
#> |Ammecr1l |  9.721498|  9.907029|  9.778428|  9.991201|  9.978755|  9.587771| 10.063778|  9.715232|  9.721717| 10.323631| 11.199652|  9.580129|  9.660760|  9.778464|  9.895509|  9.344508|  9.896869|  9.459963|  9.705875|  9.639385|  9.505102| 10.192027| 10.533754|  9.888044|
#> |Amn      |  0.000000|  4.398652|  4.802926|  5.474513|  5.608943|  5.390806|  0.000000|  1.539411|  2.652177|  0.000000|  6.006920|  0.000000|  5.308966|  4.118472|  4.413498|  0.000000|  0.000000|  0.000000|  5.480945|  4.193651|  3.798388|  5.562926|  5.092165|  5.357062|
#> |Amn1     | 10.324317| 10.370562|  9.845639|  9.861057|  8.840841|  9.157939| 10.075279|  9.444744|  9.264037|  9.709497|  6.300475|  9.612447| 10.030505|  9.414175|  9.932386|  9.211777|  9.085315|  9.152905|  8.525869|  8.986494|  9.161191|  9.073315| 10.736147|  9.381018|
#> |Amot     |  9.777482|  9.208478|  9.767862| 12.542766| 11.746755| 11.902718|  9.697512|  9.531469|  9.754372| 12.228722| 11.930041| 11.833651| 10.397455|  9.779157|  9.961128| 11.908335| 11.970097| 11.769010|  8.810641|  8.698924|  9.029949| 12.112950| 12.567338| 12.034565|
#> |Amotl1   |  9.522070| 10.198338|  7.256056| 11.293638| 11.748941| 10.654560|  6.231732|  7.410853|  7.114902| 11.887076| 11.454849| 11.002505|  8.893451|  8.496768|  9.317338| 10.812618| 11.228855| 11.024140|  8.847757|  8.582899|  8.707775| 11.815032| 12.292345| 11.234173|
#> |Amotl2   |  7.538139|  7.101209|  8.172426| 11.284778|  9.590328|  8.587643|  6.689631|  6.728571|  6.042835| 11.836940|  9.460240|  9.869199|  8.022048|  7.368093|  7.296374|  8.879769| 10.364257| 10.277043|  6.797734|  6.442407|  6.686830|  9.631151|  9.846421|  9.230496|