Farris et al transcript-level expression data

Farris et al transcript-level expression data

farrisTxSE

Format

SummarizedExperiment object with 122733 transcript rows, and 24 sample columns. Sample annotations are accessed as described for SummarizedExperiment-class objects, using SummarizedExperiment::colData(). Transcript and gene annotations are accessed using SummarizedExperiment::rowData().

The colData colnames are described as follows:

CellType

The major excitatory pyramidal cell types in the mouse hippocampus, CA1, CA2, CA3, and dentate gyrus (DG).

Compartment

The cellular compartment within each cell type: CB (cell body) is the pyramidal cell layer; DE (dendrite) is stratum radiatum of CA1, CA2, CA3, and the molecular layer of DG.

AnimalID

The number assigned to each mouse.

groupName

The sample group name associated with each CellType and Compartment.

The count data is accessed through the function assays and has the following entries:

tpm

Median-normalized log2-transformed TPM values generated by Salmon, after import using tximport::tximport(). The log2 transformation was log2(1+x).

counts

Median-normalized log2-transformed pseudocount data per-gene generated by Salmon, after import using tximport::tximport(). The log2 transformation was log2(1+x).

raw_tpm

Raw log2-transformed TPM values generated by Salmon, after import using tximport::tximport(). The log2 transformation was log2(1+x).

raw_counts

The raw log2-transformed pseudocount data per-gene generated by Salmon. The log2 transformation was log2(1+x).

The rowData contains gene annotations, which includes the relationship between transcript and gene.

transcript_id

Gencode transcript_id as defined in the GTF file.

geneSymbol

Most current Entrez gene symbol associated with the transcript_id, as of June 2018.

gene_type,transcript_type

Gencode attributes as defined in the GTF file.

has3UTR,TxHas3UTR

Boolean indicating whether the gene or transcript contained an annotated 3'UTR region defined in the Gencode GTF file.

TxHasExt3UTR,GeneHasExt3UTR

Following 3'UTR extension based upon RNA-seq data coverage, this boolean field indicates whether each transcript or gene contains a 3'UTR with or without extension by observed RNA-seq coverage.

TxDetectedByTPM

Boolean indicating whether each transcript was considered "detected" based upon TPM and count data per gene, as described in splicejam::defineDetectedTx().

Source

Farris et al publication.

Details

SummarizedExperiment object containing transcript-level expression data following Salmon quantitation, and import using tximport::tximport() at the transcript level. Data is stored as log2-transformed TPM, and log2-transformed pseudocounts from Salmon, after median normalization.

Examples

data(farrisTxSE);
## Transcript-Gene annotation DataFrame
head(SummarizedExperiment::rowData(farrisTxSE))

#> DataFrame with 6 rows and 10 columns
#>                             transcript_id    geneSymbol            gene_type
#>                               <character>   <character>          <character>
#> ENSMUST00000193812.1 ENSMUST00000193812.1 4933401J01Rik                  TEC
#> ENSMUST00000082908.1 ENSMUST00000082908.1       Gm26206                snRNA
#> ENSMUST00000162897.1 ENSMUST00000162897.1          Xkr4       protein_coding
#> ENSMUST00000159265.1 ENSMUST00000159265.1          Xkr4       protein_coding
#> ENSMUST00000070533.4 ENSMUST00000070533.4          Xkr4       protein_coding
#> ENSMUST00000192857.1 ENSMUST00000192857.1       Gm18956 processed_pseudogene
#>                           transcript_type   has3UTR TxHas3UTR TxHasExt3UTR
#>                               <character> <logical> <logical>    <logical>
#> ENSMUST00000193812.1                  TEC     FALSE     FALSE        FALSE
#> ENSMUST00000082908.1                snRNA     FALSE     FALSE        FALSE
#> ENSMUST00000162897.1 processed_transcript      TRUE     FALSE        FALSE
#> ENSMUST00000159265.1 processed_transcript      TRUE     FALSE        FALSE
#> ENSMUST00000070533.4       protein_coding      TRUE      TRUE         TRUE
#> ENSMUST00000192857.1 processed_pseudogene     FALSE     FALSE        FALSE
#>                      GeneHasExt3UTR  TxHasCDS TxDetectedByTPM
#>                           <logical> <logical>       <logical>
#> ENSMUST00000193812.1          FALSE     FALSE           FALSE
#> ENSMUST00000082908.1          FALSE     FALSE           FALSE
#> ENSMUST00000162897.1           TRUE     FALSE            TRUE
#> ENSMUST00000159265.1           TRUE     FALSE            TRUE
#> ENSMUST00000070533.4           TRUE      TRUE            TRUE
#> ENSMUST00000192857.1          FALSE     FALSE           FALSE

## Sample annotation DataFrame
head(SummarizedExperiment::colData(farrisTxSE))

#> DataFrame with 6 rows and 4 columns
#>          CellType   Region    AnimalID groupName
#>          <factor> <factor> <character>  <factor>
#> CA1CB492      CA1       CB         492    CA1_CB
#> CA1CB496      CA1       CB         496    CA1_CB
#> CA1CB502      CA1       CB         502    CA1_CB
#> CA1DE492      CA1       DE         492    CA1_DE
#> CA1DE496      CA1       DE         496    CA1_DE
#> CA1DE502      CA1       DE         502    CA1_DE

## Assay data matrix for normalized TPM
head(SummarizedExperiment::assays(farrisTxSE)[["tpm"]]);

#>                      CA1CB492  CA1CB496 CA1CB502 CA1DE492 CA1DE496 CA1DE502
#> ENSMUST00000193812.1 1.610332 0.7904088 1.158373 1.792008 1.311585 1.190829
#> ENSMUST00000082908.1 0.000000 0.0000000 0.000000 0.000000 0.000000 0.000000
#> ENSMUST00000162897.1 5.546307 5.4100021 5.704815 4.224328 5.491877 4.608420
#> ENSMUST00000159265.1 4.900158 4.8581246 5.128569 4.785710 5.218501 3.267856
#> ENSMUST00000070533.4 4.819636 4.9697168 4.962035 3.960422 4.062039 3.999063
#> ENSMUST00000192857.1 1.586775 1.3612144 1.715048 0.000000 0.000000 0.000000
#>                      CA2CB492 CA2CB496 CA2CB502 CA2DE492 CA2DE496 CA2DE502
#> ENSMUST00000193812.1 1.356880 1.122915 1.922335 1.548804 1.163878 0.000000
#> ENSMUST00000082908.1 0.000000 0.000000 0.000000 0.000000 0.000000 0.000000
#> ENSMUST00000162897.1 4.882472 5.319941 4.528681 3.287487 3.987989 4.152704
#> ENSMUST00000159265.1 4.534462 4.779240 4.552718 4.084830 3.873951 3.859247
#> ENSMUST00000070533.4 4.804406 4.889756 4.254999 3.394788 3.973354 3.841563
#> ENSMUST00000192857.1 0.000000 1.056060 0.000000 0.000000 0.000000 0.000000
#>                      CA3CB492  CA3CB496  CA3CB502 CA3DE492 CA3DE496 CA3DE502
#> ENSMUST00000193812.1 0.000000 0.9701784 0.9787647 0.000000 1.301224 0.000000
#> ENSMUST00000082908.1 0.000000 0.0000000 0.0000000 0.000000 0.000000 0.000000
#> ENSMUST00000162897.1 5.646500 5.7608985 5.6991368 4.682759 5.313276 4.654305
#> ENSMUST00000159265.1 5.506492 4.8380664 5.4546917 3.988007 4.436601 4.002717
#> ENSMUST00000070533.4 5.310857 5.2492192 5.3960733 4.334719 4.413020 4.157208
#> ENSMUST00000192857.1 0.000000 1.5506465 0.0000000 0.000000 0.000000 0.000000
#>                        DGCB492   DGCB496  DGCB502   DGDE492  DGDE496  DGDE502
#> ENSMUST00000193812.1 0.4902402 0.6670767 0.643377 0.9595546 0.000000 1.044818
#> ENSMUST00000082908.1 0.0000000 0.0000000 0.000000 0.0000000 0.000000 0.000000
#> ENSMUST00000162897.1 4.1109933 4.5889056 4.471882 3.5582606 3.565559 3.518744
#> ENSMUST00000159265.1 4.4897173 4.0607367 3.978226 2.6384302 2.929368 2.927404
#> ENSMUST00000070533.4 3.9907276 4.0903952 3.950510 3.2993651 3.718393 3.322432
#> ENSMUST00000192857.1 0.0000000 1.2529288 1.871817 0.0000000 0.000000 0.000000