farrisdata: Farris et al RNA-seq data

Supporting data for Farris et al Mouse hippocampus RNA-seq analysis workflow, utilizing the jampack package.

Details

These objects are provided as SummarizedExperiment objects as described in the Bioconductor package SummarizedExperiment.

All gene symbols were updated to the most recent annotations available at the time of publication (2019) and should be updated again for comparison with analysis results produced after that time. One option is to use Github R package "jmw86069/genejam" which relies upon Bioconductor gene annotation packages such as "org.Mm.eg.db" (mouse), used with argument: ann_lib="org.Mm.eg.db". An example command:

Ainsley_gdf <- genejam::freshenGenes3(AinsleyGenes,
   ann_lib="org.Mm.eg.db",
   intermediate="ENTREZID",
   include_source=TRUE,
   empty_rule="original")

As of October 2024, only one gene symbol was not found: "AI314180". All other gene symbols were correctly identified.

Available data

farrisGeneSE, farrisTxSE

Gene and transcript models

Gene and transcript models were defined in the corresponding Gencode GTF file available from the Gencode FTP site:

gencode.vM12.annotation.gtf.gz

Alignment to the transcriptome was performed using the FASTA sequence file also from Gencode FTP site:

gencode.vM12.transcripts.fa.gz

Gencode data is subject to Creative Commons License (Attribution, Non-commercial, No-derivatives 2.5 license), see https://creativecommons.org/licenses/by-nc-nd/2.5/.

See also